Code/CloudFunctions/store-food-tem/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-pubsub
firebase-admin
google-cloud-firestore